Improved change password form functionality and fixed login.php bugs
Added functionality to changePassword.php that checks for correct existing password, different new password, check to see if new password and confirm password are the same, added proper error messages for each case. It does not currently replace password within database.
Login page had problems with signing in to account where it allowed user to sign in with only an existing email. I added in functionality to properly check email and password and output the appropriate error message.